[Bronze I] Title: 피보나치 수 2, Time: 68 ms, Memory: 5532 KB -BaekjoonHub

+++ 백준/Bronze/2748. 피보나치 수 2/피보나치 수 2.cs
@@ -7,9 +7,9 @@ namespace Algorism1
         {
             int inputNum = Convert.ToInt32(Console.ReadLine());
 
-            decimal num = 0;
-            decimal num2 = 1;
-            decimal num3 = 1;
+            long num = 0;
+            long num2 = 1;
+            long num3 = 1;
             for (int i = 0; i < inputNum - 1; i++)
             {
                 num3 = num + num2;
